+++ arancino/ArancinoCommandExecutor.py
@@ -149,7 +149,7 @@ class ArancinoCommandExecutor:
             semver_value_libvers = semver.Version(value_libvers)
 
             # store the reserved key
-            self.__datastore.set(key_libvers, value_libvers)
+            # self.__datastore.set(key_libvers, value_libvers)
 
             # and then check if it's compatible. if the library is not compatible, disconnect the board and
             # if value_libvers not in self.compatibility_array: